pyproject.toml
[tool.poetry]
name = "dilbert"
version = "0.1.0"
description = ""
authors = ["Alphacode18 <skharban@purdue.edu>"]

[tool.poetry.scripts]
start-bot = "dilbert.main:main"

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.18.1"
slack-bolt = "^1.14.3"
python-dotenv = "^0.20.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
